third_party/blink/web_tests/css3/blending/isolation-parsing-expected.txt

PASS setProperty("auto") is "auto"
PASS setProperty("isolate") is "isolate"
PASS setProperty("") is "auto"
PASS setProperty("rubbish") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
